requirement.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
duckdb==1.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
massive==2.2.0
numpy==2.4.2
pandas==3.0.0
polygon-api-client==1.14.5
python-dateutil==2.9.0.post0
requests==2.32.5
-e git+https://github.com/dotd/scanner_minute.git@42d44917f603bc34de8cfebd311c18217e3fd7b0#egg=ScannerMinute
setuptools==82.0.0
six==1.17.0
urllib3==2.6.3
websocket-client==1.9.0
websockets==16.0
